backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
sqlalchemy[asyncio]==2.0.30
asyncpg==0.29.0
redis==5.0.4
google-generativeai==0.7.0
python-slugify==8.0.4
python-dotenv==1.0.1
apscheduler==3.10.4
